intermediary-app/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.23.0
python-dotenv>=1.0.0
loguru>=0.7.0
pyzkfp
